==> projects/TUC_MackenzieCr/analysis/03_AggregateEstimates.R <==
#' ---
#' title: "Calculate benefit and aggregate"
#' author: "Abbey Camaclang"
#' date: "27 Mar 2024"
#' output: github_document
#' ---
#'   
#' Reads compiled data (.csv files) generated by *CompileData.R*, 
#' counts the number of estimates for each Biodiversity feature,
#' calculates the Benefit then aggregates (averages) the estimates.
#' Also aggregates estimates of probability of success of the project.
#'  
#+ warning = FALSE, message = FALSE
# Load packages
library(here)
library(tidyverse)
library(naniar)

# Specify input and output filepaths
derived <- here("projects", "TUC_MackenzieCr", "data", "initial", "derived") 
results <- here("projects", "TUC_MackenzieCr", "analysis", "initial")

# Load custom functions
code <- here("R")
source(paste0(code, "/import.R")) 

#' Probability of project success
feasdata <- read.csv(paste(derived, "/Feas_estimates.csv", sep = ""))

feasdata <- feasdata %>%
  filter(!Value == "") # removes expert notes/comments

write.csv (feasdata, 
           paste(derived, "/Feas_estimates_tidy.csv", sep = ""), 
           row.names = FALSE)

feasdata$Value <- as.numeric(feasdata$Value) # also coerces blanks to NAs

# Count number of estimates
temp.feas <- na.omit(feasdata)
feas.count <- table(temp.feas$Estimate)

write.csv(feas.count, 
          file = paste(results, "/Feas_estimatecounts.csv", sep = ""), 
          row.names = FALSE)
# Aggregate estimates
feasibility <- aggregate(Value ~ Estimate, data = feasdata,
                         FUN = mean, na.action = na.omit) %>%
  mutate(Mean = Value/100)
bestfeas <- feasibility$Mean[which(feasibility$Estimate == "BEST GUESS")]

write.csv(feasibility, 
          file = paste(results, "/Feasibility_avg.csv", sep = ""), 
          row.names = FALSE)


#' Species at risk

sarclean <- read.csv(paste(derived, "/SAR_Estimates.csv", sep = ""))

# Count the number of estimates per species x scenario
# Use custom function to reorganize data and determine number of estimates
sar <- count(sarclean)

sar.table <- sar$table
sar.long <- sar$tidy
sar.count <- sar$count

write.csv(sar.long, 
          file = paste(derived, "/SAR_Estimates_tidy.csv", sep = ""), 
          row.names = FALSE)
write.csv(sar.count, 
          file = paste(results, "/SAR_estimatecounts.csv", sep = ""))


# Calculate benefit
# If Counterfactual and Action estimates are both coded as "B"
# recode using a dummy value to calculate benefit
# then convert to NA for plotting purposes
sar.temp <- sar.table
sar.temp$Counterfactual[which(str_detect(sar.temp$Counterfactual, "B") == 1)] <- 1000
sar.temp$Action[which(str_detect(sar.temp$Action, "B") == 1)] <- 1000
sar.temp$Counterfactual <- as.numeric(sar.temp$Counterfactual)
sar.temp$Action <- as.numeric(sar.temp$Action)

sar.ben <- sar.temp %>%
  mutate(Benefit = Action - Counterfactual) %>%
  mutate(Counterfactual = na_if(Counterfactual, 1000)) %>%
  select(-Action)

# Aggregate estimates across experts
# Use custom function to aggregate estimates, calculate expected benefit 
# and performance, and generate table of results
# Requires estimated mean feasibility ('bestfeas') from above
sar.results <- calculate(sar.ben, sar.count, bestfeas)

write.csv(sar.results, 
          file = paste(results, "/SAR_results.csv", sep = ""), 
          row.names = FALSE)


#' Functional groups by ecotype and taxon

grpclean <- read.csv(paste(derived, "/FuncGrp_Estimates.csv", sep = ""))

# Count number of expert estimates per species x scenario
grp <- count(grpclean)

grp.table <- grp$table
grp.long <- grp$tidy
grp.count <- grp$count

write.csv(grp.long, 
          file = paste(derived, "/FuncGrp_Estimates_tidy.csv", sep = ""), 
          row.names = FALSE)
write.csv(grp.count, 
          file = paste(results, "/FuncGrp_estimatecounts.csv", sep = ""))

# Calculate benefit
# Fix Counterfactual estimates coded as "B"
grp.temp <- grp.table
grp.temp$Counterfactual[which(str_detect(grp.temp$Counterfactual, "B") == 1)] <- 1000
grp.temp$Action[which(str_detect(grp.temp$Action, "B") == 1)] <- 1000
grp.temp$Counterfactual <- as.numeric(grp.temp$Counterfactual)
grp.temp$Action <- as.numeric(grp.temp$Action)

# Get benefit value then convert 'B' Counterfactual estimates to NA
grp.ben <- grp.temp %>%
  mutate(Benefit = Action - Counterfactual) %>%
  mutate(Counterfactual = na_if(Counterfactual, 1000)) %>%
  select(-Action)

# Aggregate estimates across experts
grp.results <- calculate(grp.ben, grp.count, bestfeas)

grp.results <- grp.results %>%
  separate(Biodiversity, c("Ecotype", "Taxa", "FuncGroup"), sep = "\\n")

write.csv(grp.results, 
          file = paste(results, "/FuncGrp_results.csv", sep = ""), 
          row.names = FALSE)


#' Ecotypes

ecoclean <- read.csv(paste(derived, "/Ecotype_Estimates.csv", sep = ""))

# Count number of expert estimates per species x scenario
eco <- count(ecoclean)

eco.table <- eco$table
eco.long <- eco$tidy
eco.count <- eco$count

write.csv(eco.long, 
          file = paste(derived, "/Ecotype_Estimates_tidy.csv", sep = ""), 
          row.names = FALSE)
write.csv(eco.count, 
          file = paste(results, "/Ecotype_estimatecounts.csv", sep = ""))

# Calculate benefit
# Fix Counterfactual estimates coded as "B"
eco.temp <- eco.table
eco.temp$Counterfactual[which(str_detect(eco.temp$Counterfactual, "B") == 1)] <- 1000
eco.temp$Action[which(str_detect(eco.temp$Action, "B") == 1)] <- 1000
eco.temp$Counterfactual <- as.numeric(eco.temp$Counterfactual)
eco.temp$Action <- as.numeric(eco.temp$Action)

# Get benefit value then convert 'B' Counterfactual estimates to NA
eco.ben <- eco.temp %>%
  mutate(Benefit = Action - Counterfactual) %>%
  mutate(Counterfactual = na_if(Counterfactual, 1000)) %>%
  select(-Action)

# Aggregate estimates across experts
eco.results <- calculate(eco.ben, eco.count, bestfeas)

write.csv(eco.results, file = paste(results, "/Ecotype_results.csv", sep = ""), row.names = FALSE)
